python/requirements.txt
Flask
Flask-JSON
mysql-connector-python